+++ appimage_recipe.sh
@@ -53,14 +53,14 @@ rm -fr -v ${MY_PYDIR}/distutils
 echo "Copy built-in script patch from '${PYDIR}' to '${MY_PYDIR}' ..."
 cd ${PYDIR}
 for f in *; do
-    if [ ${f} == "__pycache__" ] || [ ${f} == "test" ] \
-        || [ ${f} == "venv" ] || [ ${f} == "idlelib" ]; then continue; fi
+    if [[ ${f} == "__pycache__" ]] || [[ ${f} == "test" ]] \
+        || [[ ${f} == "venv" ]] || [[ ${f} == "idlelib" ]]; then continue; fi
 
     if [[ ${f} == *.py ]]; then
         cp -n -v ${f} ${MY_PYDIR}
     fi
 
-    if [ ! -d ${f} ]; then continue; fi
+    if [[ ! -d ${f} ]]; then continue; fi
 
     echo "Create '${MY_PYDIR}/${f}'"
     mkdir -p ${MY_PYDIR}/${f}
@@ -69,13 +69,13 @@ for f in *; do
     cd ${f}
 
     for sub_f in *; do
-        if [ ${sub_f} == "__pycache__" ]; then continue; fi
+        if [[ ${sub_f} == "__pycache__" ]]; then continue; fi
 
         if [[ ${sub_f} == *.py ]]; then
             cp -n -v ${sub_f} ${MY_PYDIR}/${f}
         fi
 
-        if [ ! -d ${sub_f} ]; then continue; fi
+        if [[ ! -d ${sub_f} ]]; then continue; fi
 
         echo "Create '${MY_PYDIR}/${f}/${sub_f}'"
         mkdir -p ${MY_PYDIR}/${f}/${sub_f}
@@ -130,6 +130,7 @@ cat > ${LOWERAPP}.desktop <<EOF
 Name=${APP}
 Exec=${LOWERAPP}
 Type=Application
+Categories=Development;Education;
 Icon=${LOWERAPP}
 StartupNotify=true
 Comment=Open Source Planar Linkage Mechanism Simulation and Dimensional Synthesis System.